requirements.txt
streamlit
streamlit-drawable-canvas
numpy
opencv-python
pillow
pandas
